refactor(view): P6-A3 final cut — extract claude_bundle.cpp
Phase 6 A3 (design_import.cpp split), final cut. The pulp #468
runtime-import subsystem — the largest single cluster in design_import.cpp
— relocated into core/view/src/claude_bundle.cpp (2,459 lines):

  * Claude Design bundle envelope parsing (asset extraction, zip /
    base64 payloads)
  * parse_claude_html_with_runtime — the offline harness that boots a
    ScriptEngine/View/WidgetBridge, evaluates the bundle, and walks the
    materialized DOM into a DesignIR
  * the shared pre-payload shim + payload pipeline used by both the
    offline harness and the live-engine path
  * WidgetBridge::evaluate_claude_bundle_in_live_engine — the live
    runtime path reusing the caller's mounted reconciler

Definitions only; declarations stay in design_import.hpp (and
widget_bridge.hpp for the WidgetBridge member). The cluster reuses the
existing design_import_internal.hpp for design_source_vendor_key.

design_import.cpp: 3,324 → 913 (-2,411). Across all three A3 cuts
(design_tokens + design_codegen + claude_bundle), design_import.cpp is
down 4,691 → 913 (-80%) — now just source helpers, classname
extraction, audio-widget detection, JSON parsing, IR-from-JSON, and the
source adapters.

Verified: pulp-view builds clean; all 4 runtime-import test suites pass
(design-import 30, react-runtime 21, claude-runtime 4,
widget-bridge-runtime-import 12).
